LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/dedicated-worker-cache-storage.https-expected.txt


PASS coep-none coep-none corp-cross-origin 
PASS coep-none coep-none corp-undefined 
PASS coep-none coep-require-corp corp-cross-origin 
FAIL coep-none coep-require-corp corp-undefined assert_equals: expected "failure" but got "success"
FAIL coep-require-corp coep-none corp-cross-origin assert_equals: expected "error" but got "success"
FAIL coep-require-corp coep-none corp-undefined assert_equals: expected "error" but got "success"
PASS coep-require-corp coep-require-corp corp-cross-origin 
FAIL coep-require-corp coep-require-corp corp-undefined assert_equals: expected "failure" but got "success"
